Huge WarehouseHandler refactor, including async backoff
